Feature: Adding Nodes to the Open Vector
In the code, nodes are implemented with the type vector<int>, and should have the form {x, y, g, h} for ints x, y, g, and h.
Also, the open list (openodes) will be implemented as a C++ vector (of type vector<vector<int>>).
The goal is to wrie a helper function for the A*Search algorithm which will add nodes to the open vector and mark them as visited in the grid