Bug Fix: initialization type error | Feature: While loop for the A* algorithm
corrected initialization by different types in Search function
implemented the primary while loop in the algorithm which sorts the open list using f-value, and using the node with the lowest f-value as the next node in the search. This process continues until the goal node has been found or the open list runs out of nodes to use for searching.

Modified the HelperFunctions.h file and added a new CellSort function